+++ source/core/scene/animation/animation.hpp
@@ -9,9 +9,10 @@ namespace scene {
 namespace animation {
 
 class Animation {
+
 public:
 
-    void init(size_t count);
+	void init(uint32_t count);
 
 	void push_back(const entity::Keyframe& keyframe);
 
@@ -26,7 +27,7 @@ private:
 
     float current_time_;
 
-    size_t current_frame_;
+	uint32_t current_frame_;
 
 	std::vector<entity::Keyframe> keyframes_;
